Upgraded finnhub to v3 framework (#2695)
