+++ .changeset/silent-moose-camp.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/finnhub-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
