+++ .changeset/sour-roses-play.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/market-closure-adapter': minor
+---
+
+Updated implementation to account for Finnhub v3 upgrade
