+++ packages/core/legos/tsconfig.json
@@ -230,9 +230,6 @@
       "path": "../../sources/finnhub"
     },
     {
-      "path": "../../sources/finnhub-test"
-    },
-    {
       "path": "../../sources/fixer"
     },
     {
