--- packages/sources/finnhub-test/CHANGELOG.md
@@ -1,26 +1,0 @@
-# @chainlink/finnhub-test-adapter
-
-## 1.0.2
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.0.0
-
-### Major Changes
-
-- f004dfef7: Add finnhub v3 EA
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
