--- packages/sources/finnhub-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as quote } from './quote'
