--- packages/sources/finnhub/schemas/env.json
@@ -1,16 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/finnhub-adapter.json",
-  "title": "Chainlink External Adapter for Finnhub",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
