+++ packages/sources/finnhub/src/config/index.ts
@@ -1,14 +1,15 @@
-import { Requester } from '@chainlink/ea-bootstrap'
-import { Config } from '@chainlink/ea-bootstrap'
+import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
 
-export const NAME = 'FINNHUB'
-
-export const DEFAULT_ENDPOINT = 'quote'
-export const DEFAULT_BASE_URL = 'https://finnhub.io/api/v1/'
-
-export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix, true)
-  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
-  config.defaultEndpoint = DEFAULT_ENDPOINT
-  return config
-}
+export const config = new AdapterConfig({
+  API_ENDPOINT: {
+    description: 'The HTTP URL to retrieve data from',
+    type: 'string',
+    default: 'https://finnhub.io/api/v1',
+  },
+  API_KEY: {
+    description: 'A Finnhub API key ',
+    type: 'string',
+    required: true,
+    sensitive: true,
+  },
+})
