--- packages/sources/finnhub/src/config/limits.json
@@ -1,12 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1m": 60
-    },
-    "all-in-one": {
-      "rateLimit1m": 300,
-      "note": "limit is for market data, not fundamental data"
-    }
-  },
-  "ws": {}
-}
