+++ packages/sources/finnhub/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as QuoteInputParameters } from './quote'
-
-export type TInputParameters = QuoteInputParameters
-
-export * as quote from './quote'
+export { endpoint as quote } from './quote'
