+++ packages/sources/finnhub/test-payload.json
@@ -1,5 +1,5 @@
 {
     "requests": [{
-        "base": "XAU"
+        "base": "EUR"
     }]
 }
