+++ packages/sources/finnhub/test/integration/adapter.test.ts
@@ -1,66 +1,65 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
 import * as process from 'process'
-import { server as startServer } from '../../src'
-import { mockResponseFailure, mockResponseSuccess } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
-import { SuperTest, Test } from 'supertest'
+import { AddressInfo } from 'net'
+import request, { SuperTest, Test } from 'supertest'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { AdapterRequestBody, sleep } from '@chainlink/external-adapter-framework/util'
+import { mockResponseSuccess } from './fixtures'
+import { setEnvVariables, createAdapter } from './setup'
 
-describe('execute', () => {
-  const id = '1'
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
-
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-    API_KEY: process.env.API_KEY || 'fake-api-key',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('rate api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'EUR',
-      },
-    }
-
-    it('should return success', async () => {
-      mockResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
+describe('rest', () => {
+  jest.setTimeout(10000)
+  let spy: jest.SpyInstance
+  beforeAll(async () => {
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
   })
 
-  describe('rate api with invalid base', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'NON-EXISTING',
-      },
-    }
+  afterAll((done) => {
+    spy.mockRestore()
+    done()
+  })
 
-    it('should return failure', async () => {
-      mockResponseFailure()
+  let fastify: ServerInstance | undefined
+  let req: SuperTest<Test>
 
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(400)
+  const data: AdapterRequestBody = {
+    data: {
+      base: 'EUR',
+    },
+  }
+
+  let oldEnv: NodeJS.ProcessEnv
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['CACHE_MAX_AGE'] = '5000'
+    process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
+    process.env['METRICS_ENABLED'] = 'false'
+    process.env['API_KEY'] = 'fake-api-key'
+    fastify = await expose(createAdapter())
+    req = request(`http://localhost:${(fastify?.server.address() as AddressInfo).port}`)
+    mockResponseSuccess()
+    // Send initial request to start background execute
+    await req.post('/').send(data)
+    await sleep(5000)
+  })
+
+  afterAll((done) => {
+    setEnvVariables(oldEnv)
+    fastify?.close(done())
+  })
+
+  describe('quote endpoint', () => {
+    it('should return success', async () => {
+      const makeRequest = () =>
+        req
+          .post('/')
+          .send(data)
+          .set('Accept', '*/*')
+          .set('Content-Type', 'application/json')
+          .expect('Content-Type', /json/)
+
+      const response = await makeRequest()
       expect(response.body).toMatchSnapshot()
-    })
+    }, 30000)
   })
 })
