+++ packages/sources/finnhub/test/integration/fixtures.ts
@@ -29,20 +29,3 @@ export const mockResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
-
-export const mockResponseFailure = (): nock.Scope =>
-  nock('https://finnhub.io/api/v1', {
-    encodedQueryParams: true,
-  })
-    .get('/quote')
-    .query({ token: 'fake-api-key', symbol: 'NON-EXISTING' })
-    .reply(200, () => ({ c: 0, d: null, dp: null, h: 0, l: 0, o: 0, pc: 0, t: 0 }), [
-      'Content-Type',
-      'application/json',
-      'Connection',
-      'close',
-      'Vary',
-      'Accept-Encoding',
-      'Vary',
-      'Origin',
-    ])
