+++ packages/tsconfig.json
@@ -342,9 +342,6 @@
       "path": "./sources/finnhub"
     },
     {
-      "path": "./sources/finnhub-test"
-    },
-    {
       "path": "./sources/fixer"
     },
     {
